Introduce StkTru token (#451)
* Introduce StkTru contract

* fix contract lint

* Rework staking contract

* Add oracle setter

* Stake - no claimable rewards

* Rework contract

* Add tests

* Add warning

* add withdraw function

+++ contracts/governance/interface/IVoteToken.sol
@@ -1,10 +1,20 @@
-// SPDX-License-Identifier: MIT 
+// SPDX-License-Identifier: MIT
 
 pragma solidity ^0.6.10;
 
 interface IVoteToken {
     function delegate(address delegatee) external;
-    function delegateBySig(address delegatee, uint nonce, uint expiry, uint8 v, bytes32 r, bytes32 s) external;
+
+    function delegateBySig(
+        address delegatee,
+        uint256 nonce,
+        uint256 expiry,
+        uint8 v,
+        bytes32 r,
+        bytes32 s
+    ) external;
+
     function getCurrentVotes(address account) external view returns (uint96);
-    function getPriorVotes(address account, uint blockNumber) external view returns (uint96);
+
+    function getPriorVotes(address account, uint256 blockNumber) external view returns (uint96);
 }
